+++ entity/NPC_0.java
@@ -1,0 +1,13 @@
+package entity;
+
+import main.GamePanel;
+
+public class NPC_0 extends Entity{
+
+    public NPC_0(GamePanel gp){
+       //super(gp);
+        
+        direction = "down";
+        speed = 2;
+    }
+}