+++ object/OBJ_0.java
@@ -1,0 +1,17 @@
+package object;
+
+import java.io.IOException;
+
+import javax.imageio.ImageIO;
+
+public class OBJ_0 extends SuperObject{
+
+    public OBJ_0(){
+        name = "Objeto 0";
+        try{
+            image = ImageIO.read(getClass().getResourceAsStream("../res/objects/object_0.png"));
+        }catch(IOException e){
+            e.printStackTrace();
+        }
+    }
+}